Move hamiltonian class (#1534)
* move

* black

* move tests

* Update .github/CHANGELOG.md

* black

* fix codefactor

* add pylint to unrelated code line

* black

* black

* fix qchem import

* minor change

* Update pennylane/ops/qubit/hamiltonian.py


* Update .github/CHANGELOG.md


* add deprecation warning and docs entry

* black

* remove deprecation again
